requirements.txt
# 核心依赖
openai-whisper==20240930
openai>=1.35.0
ffmpeg-python==0.2.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.0.0
click>=8.1.0
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# 开发工具
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.10.0
pytest-asyncio>=0.23.0